version 0.0.2 (pre-alpha)
NEW:
* Initial release of stats.php file, which reads minecraft stats JSON
files from the server. It has been changed a bit, and is now acceptable
for testing purposes. It currently requests a player name to lookup, and
it returns a neatly formatted PHP array of the contents. It doesn't look
fancy at all, and for those without prior knowledge of the minecraft
stats, item ID, etc. may not understand the results.

FIXED:
* none